It's a bad idea to test code in 'index.js' because it interferes with GHA. Move createRelease back to the wrapper
